+ A missing check for an old version backup has been added inside the “overrideOldVersionFile” method.

+++ src/Installer.php
@@ -47,9 +47,10 @@ class Installer extends LibraryInstaller {
 	 */
 	private function overrideOldVersionFile(){
 		$oldVersionFilePath = $this->ddToolsDeprecatedPath.$this->ddToolsClassFileName;
+		$oldVersionBackupPath = $this->ddToolsDeprecatedPath.$this->ddToolsOldVersionBackupFileName;
 		
-		//Check if an old version of the library is in assets/snippets
-		if(is_file($oldVersionFilePath)){
+		//Check if an old version of the library is in assets/snippets and its backup doesn't exist
+		if(is_file($oldVersionFilePath) && !is_file($oldVersionBackupPath)){
 			//Backup of the old version
 			file_put_contents(
 				$this->ddToolsDeprecatedPath.$this->ddToolsOldVersionBackupFileName,
